refactor(sqlite): extract mixins from dialect to dedicated files
- Split monolithic mixins.py (1952 lines) into 14 individual files under mixins/
- Remove duplicate method implementations from dialect.py (1276→706 lines)
- Make SQLite mixins standalone (no base class inheritance, MySQL pattern)
- Fix MRO ordering: SQLite-specific mixins before generic ones in declaration
- Move shared constants (_SUGGESTION_*, _COLLATION_NAME_RE) to owning mixin files
- Fix ruff issues (unused imports, TYPE_CHECKING forward references)
- Update test method names from _handle_* to format_*
