+++ src/rhosocial/activerecord/backend/impl/sqlite/mixins/__init__.py
@@ -1,0 +1,39 @@
+# src/rhosocial/activerecord/backend/impl/sqlite/mixins/__init__.py
+"""
+SQLite-specific mixin implementations.
+
+This package provides mixin classes that implement SQLite-specific
+features for the ActiveRecord ORM.
+"""
+
+from .extension import SQLiteExtensionMixin
+from .virtual_table import SQLiteVirtualTableMixin
+from .pragma_mixin import SQLitePragmaMixin
+from .reindex import SQLiteReindexMixin
+from .introspection import SQLiteIntrospectionCapabilityMixin
+from .identifier import SQLiteIdentifierMixin
+from .datetime import SQLiteDateTimeMixin
+from .ddl_column import SQLiteDDLColumnMixin
+from .dml import SQLiteDMLMixin
+from .set_operation import SQLiteSetOperationMixin
+from .view import SQLiteViewMixin
+from .trigger import SQLiteTriggerMixin
+from .transaction import SQLiteTransactionMixin
+from .function import SQLiteFunctionMixin
+
+__all__ = [
+    'SQLiteExtensionMixin',
+    'SQLiteVirtualTableMixin',
+    'SQLitePragmaMixin',
+    'SQLiteReindexMixin',
+    'SQLiteIntrospectionCapabilityMixin',
+    'SQLiteIdentifierMixin',
+    'SQLiteDateTimeMixin',
+    'SQLiteDDLColumnMixin',
+    'SQLiteDMLMixin',
+    'SQLiteSetOperationMixin',
+    'SQLiteViewMixin',
+    'SQLiteTriggerMixin',
+    'SQLiteTransactionMixin',
+    'SQLiteFunctionMixin',
+]
