+++ tests/rhosocial/activerecord_test/feature/backend/sqlite/test_sqlite_dialect_security.py
@@ -68,8 +68,8 @@ def test_sqlite_format_default_constraint_string_escaping(dialect):
     )
 
     sql, params = dialect.format_default_constraint(constraint)
-    assert "test''s value" in sql
-    assert "'; DROP" not in sql
+    assert sql == " DEFAULT ?"
+    assert params == ("test's value",)
 
 
 def test_sqlite_format_storage_options_string_escaping(dialect):
